arch-system: add backup-system package
This implements fully automated, hourly backups of my /home via
borgbackup, btrfs and systemd timers.

Note that `setup_borgbackup`, a related function that is part of my
dotfiles, is needed to setup backups for the first time, as well as a
call to `borg init` with the proper environment variables.