arch-system: fix .install and run it
Before, base-system.install wasn't actually run, since it wasn't
specified in `install=` in the PKGBUILD anywhere. Also run it both in
the post-upgrade and post-install steps.

`--global` was removed since it caused errors, it's unclear what exactly
it does in the first place, but other official PKGBUILDs did it this way..